feat(mstsgu): extract PacketIo from the WebSocket gateway client
Move the HTTPS WebSocket transport, auth upgrade loop, and byte
read/write helpers out of lib.rs so handshake/tunnel/channel
sequencing stays separate from I/O.

GwClient write-side shutdown now closes the outbound sender and
waits for the worker so a local EOF can finish. Channel-close
PDUs are not added.

This does not include dual-HTTP, RPCH, proxies, or certificate
policy APIs.
